вынес тип RouterPaths отдельно, испр логику работы с активной ссылкой
