+++ packages/client/src/app.types.ts
@@ -1,0 +1,13 @@
+export enum RouterPaths {
+  MAIN = '/',
+  HOW_TO = '/how-to',
+  REGISTER = '/sign-up',
+  LOGIN = '/login',
+  PROFILE = '/profile',
+  GAME = '/game',
+  LEADERBOARD = '/leaderboard',
+  FORUM = '/forum',
+  NEW_THREAD = '/forum/new',
+  ERROR_500 = '/500',
+  ERROR_404 = '/404',
+}
