+++ packages/client/src/components/side-menu/side-menu.module.pcss
@@ -22,15 +22,18 @@
   margin: 0;
 }
 
-.mainLinksWrap {
-  display: flex;
-  flex-flow: column;
-  gap: 24px;
-}
-
 .delimiter {
   height: 1px;
   background-color: var(--color-dark-primary, #0f101b);
+}
+
+.howToText {
+  text-align: center;
+  color: var(--color-dark-primary, #0f101b);
+
+  &[data-expanded='true'] {
+    text-align: start;
+  }
 }
 
 .menuToggler {
