+++ packages/client/src/components/side-menu/side-menu.tsx
@@ -2,20 +2,21 @@ import { FC, useState } from 'react';
 import styles from './side-menu.module.pcss';
 import cn from 'classnames';
 import { NavigationLink } from 'components/navigation-link';
-import { Link } from 'react-router-dom';
+import { Link, NavLink } from 'react-router-dom';
 import { ProfileLink } from 'components/profile-link';
 import { Footer } from 'components/footer';
 import { ReactComponent as Logo } from 'assets/logo.svg';
 import { ReactComponent as LogoNarrow } from 'assets/logo-narrow.svg';
 import { Icon } from 'components/icon';
+import { RouterPaths } from 'src/app.types';
+
 import { LINKS, EXIT, PROFILE, SOCIAL_LINKS } from 'src/mock/side-menu-links';
 
 // пока что хардкод...
 const LOGO_COLOR = '#6644ec';
 
 export const SideMenu: FC = () => {
-  const [activeLink, setActiveLink] = useState(LINKS[0].link);
-  const [expanded, setExpanded] = useState(true);
+  const [expanded, setExpanded] = useState<boolean>(true);
 
   return (
     <section
@@ -30,7 +31,7 @@ export const SideMenu: FC = () => {
           ['w-1']: !expanded,
         })}
       >
-        <Link to="/" onClick={() => setActiveLink('/')}>
+        <Link to={RouterPaths.MAIN}>
           <div className={styles.logo}>
             {expanded ? (
               <Logo width="100%" height="100%" fill={LOGO_COLOR} />
@@ -40,39 +41,48 @@ export const SideMenu: FC = () => {
           </div>
         </Link>
 
-        <div className={styles.mainLinksWrap}>
-          {expanded && <p className={cn(styles.textMenu, 'text-menu')}>Меню</p>}
-          {LINKS.map(({ iconType, text, link, informer }) => (
-            <Link key={iconType} to={link} onClick={() => setActiveLink(link)}>
-              <NavigationLink
-                text={text}
-                iconType={iconType}
-                active={link === activeLink}
-                expanded={expanded}
-                informer={informer}
-              />
-            </Link>
-          ))}
-        </div>
+        {expanded && <p className={cn(styles.textMenu, 'text-menu')}>Меню</p>}
+        {LINKS.map(({ iconType, text, link, informer }) => (
+          <NavLink key={iconType} to={link}>
+            {({ isActive }) => {
+              return (
+                <NavigationLink
+                  text={text}
+                  iconType={iconType}
+                  active={isActive}
+                  expanded={expanded}
+                  informer={informer}
+                />
+              );
+            }}
+          </NavLink>
+        ))}
 
         {expanded && <div className={styles.delimiter} />}
 
-        <Link to="profile" onClick={() => setActiveLink('/profile')}>
-          <ProfileLink {...PROFILE} expanded={expanded} />
+        <NavLink to={PROFILE.link}>
+          {({ isActive }) => {
+            return (
+              <ProfileLink {...PROFILE} active={isActive} expanded={expanded} />
+            );
+          }}
+        </NavLink>
+
+        {expanded && <div className={styles.delimiter} />}
+
+        <Link
+          to={RouterPaths.HOW_TO}
+          className="text-main"
+          state={{ fromOwnHost: true }}
+        >
+          <p
+            className={cn(styles.howToText, 'text-menu')}
+            data-expanded={expanded}
+          >
+            {expanded ? 'Как играть в Fancy Colors?' : '?'}
+          </p>
         </Link>
 
-        {expanded && (
-          <>
-            <div className={styles.delimiter} />
-            <Link
-              to="/how-to"
-              className="text-main"
-              state={{ fromOwnHost: true }}
-            >
-              Как играть в Fancy Colors?
-            </Link>
-          </>
-        )}
         <button
           type="button"
           className={cn(styles.menuToggler, { [styles.expanded]: expanded })}
@@ -83,6 +93,7 @@ export const SideMenu: FC = () => {
       </nav>
       <div>
         <button
+          /* onClick={логика выхода...} */
           className={cn(styles.exit, {
             ['w-3']: expanded,
             ['w-1']: !expanded,
@@ -91,7 +102,7 @@ export const SideMenu: FC = () => {
           <NavigationLink
             text={EXIT.text}
             iconType={EXIT.iconType}
-            active={EXIT.link === activeLink}
+            active={false}
             expanded={expanded}
           />
         </button>
