+++ packages/client/src/mock/side-menu-links.ts
@@ -1,5 +1,4 @@
-import { RouterPaths } from 'src/app';
-
+import { RouterPaths } from 'src/app.types';
 export type LinkType = {
   iconType: IconType;
   link: RouterPaths;
